+++ src/validators/DVR.ts
@@ -122,17 +122,12 @@ export class DVR implements ValidationPluginInterface {
 
   handleAsyncPasses(field, resolve) {
     field.setValidationAsyncData(true);
-    field.showAsyncErrors();
     resolve();
   }
 
   handleAsyncFails(field, validation, resolve) {
-    field.setValidationAsyncData(
-      false,
-      _.head(validation.errors.get(field.path))
-    );
+    field.setValidationAsyncData(false, _.head(validation.errors.get(field.path)));
     this.executeAsyncValidation(field);
-    field.showAsyncErrors();
     resolve();
   }