+++ src/validators/YUP.ts
@@ -68,14 +68,12 @@ class YUP implements ValidationPluginInterface {
 
   handleAsyncPasses(field, resolve) {
     field.setValidationAsyncData(true);
-    field.showAsyncErrors();
     resolve();
   }
 
   handleAsyncFails(field, resolve, error) {
     field.setValidationAsyncData(false, error.errors[0]);
     this.executeAsyncValidation(field);
-    field.showAsyncErrors();
     resolve();
   }